singleLinkedList

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,12 +10,12 @@
 size_t print_list(const list_t *h)
 {
 const list_t *tmp;
-size_t i;
+unsigned int i;
 
 tmp = h;
 for (i = 0; tmp; i++)
 {
-printf("[%lu] %s\n", tmp->len, tmp->str);
+printf("[%u] %s\n", tmp->len, tmp->str);
 tmp = tmp->next;
 }
 return (i);
